separate the concepts of vif and hardware address
In some kernel versions, HWSIM_ATTR_ADDR_TRANSMITTER is expected
to be an unchanging cookie related to the radio itself, while in
others it is the virtual interface address (i.e. what goes over
the air).

In order to more fully support the former, store a separate
hwaddr field, and always use that for reporting status and sending
cloned frames.

This is not yet a full fix for that: on startup, we should also
query the current vifs and hardware addr so that the receiver
mapping is also known.